pyproject.toml
[project]
name = "ai-job-recommender-system-using-local-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apify-client>=2.3.0",
    "mcp[cli]>=1.23.1",
    "ollama>=0.6.1",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.1",
]
